refactor: merge CheckAppStartUtils into OnboardingUtils
- Move AppStart, AppStartResult, checkAppStart into OnboardingUtils.kt
- Make checkAppStart private (only entry point is onboardIfNeeded)
- Delete CheckAppStartUtils.kt
- Rename CheckAppStartTest → AppStartTest
