+++ lib/src/main/java/de/lemke/commonutils/OnboardingUtils.kt
@@ -23,9 +23,89 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Build.VERSION.SDK_INT
 import android.os.Build.VERSION_CODES.UPSIDE_DOWN_CAKE
+import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import de.lemke.commonutils.data.commonUtilsSettings
 import de.lemke.commonutils.ui.activity.CommonUtilsOOBEActivity
+
+// ── AppStart ─────────────────────────────────────────────────────────────────
+
+/** Result category of an app launch relative to the previously recorded version. */
+enum class AppStartResult { FIRST_TIME, FIRST_TIME_VERSION, NORMAL }
+
+/** Snapshot of version and TOS state captured at app launch. */
+class AppStart(
+    val result: AppStartResult,
+    val versionCode: Int,
+    val versionName: String,
+    val lastVersionCode: Int,
+    val lastVersionName: String,
+    val tosVersion: Int,
+    val acceptedTosVersion: Int,
+) {
+    /** `true` if no previous installation was recorded. */
+    val isFirstTime get() = lastVersionCode == -1
+
+    /** `true` if the app was upgraded since the last launch. */
+    val isFirstTimeVersion get() = lastVersionCode < versionCode
+
+    /** `true` if the user has accepted the current TOS version. */
+    val tosAccepted get() = acceptedTosVersion >= tosVersion
+
+    /** `true` if OOBE should be shown (first install or TOS not accepted). */
+    val shouldShowOOBE get() = isFirstTime || !tosAccepted
+
+    /** `true` if [threshold] falls within the range of version codes updated across on this launch. */
+    fun versionThresholdPassed(threshold: Int) = threshold in lastVersionCode..<versionCode
+
+    override fun toString(): String =
+        "AppStart(result=$result, versionCode=$versionCode, versionName='$versionName', " +
+            "lastVersionCode=$lastVersionCode, lastVersionName='$lastVersionName', " +
+            "tosVersion=$tosVersion, acceptedTosVersion=$acceptedTosVersion)"
+}
+
+private const val TAG = "AppStart"
+
+/** Checks whether this is the first run, a version upgrade, or a normal start; persists version info for next launch. */
+private fun AppCompatActivity.checkAppStart(
+    versionCode: Int,
+    versionName: String,
+): AppStart {
+    val lastVersionCode = commonUtilsSettings.lastVersionCode
+    val lastVersionName = commonUtilsSettings.lastVersionName
+    commonUtilsSettings.lastVersionCode = versionCode
+    commonUtilsSettings.lastVersionName = versionName
+    val tosVersion = resources.getInteger(R.integer.commonutils_tos_version)
+    val acceptedTosVersion = commonUtilsSettings.acceptedTosVersion
+    val result =
+        when {
+            lastVersionCode == -1 -> {
+                AppStartResult.FIRST_TIME
+            }
+
+            lastVersionCode < versionCode -> {
+                AppStartResult.FIRST_TIME_VERSION
+            }
+
+            lastVersionCode > versionCode -> {
+                Log.w(TAG, "Current version code ($versionCode) is less then the one recognized on last startup ($lastVersionCode). ")
+                Log.w(TAG, "Defensively assuming normal app start.")
+                AppStartResult.NORMAL
+            }
+
+            else -> {
+                AppStartResult.NORMAL
+            }
+        }
+    return AppStart(result, versionCode, versionName, lastVersionCode, lastVersionName, tosVersion, acceptedTosVersion).apply {
+        Log.d(TAG, this.toString())
+        if (result == AppStartResult.FIRST_TIME_VERSION && !tosAccepted) {
+            CommonUtilsOOBEActivity.tosChanged = true
+        }
+    }
+}
+
+// ── Onboarding chain ──────────────────────────────────────────────────────────
 
 /** Intent extra marking an activity launched as a step of the onboarding chain. */
 const val EXTRA_ONBOARDING_STEP = "commonUtilsOnboardingStep"
